+++ src/main/java/nextstep/ladder/model/Ladder.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import java.util.List;
 import java.util.stream.Collectors;
