+++ src/test/java/nextstep/ladder/LineTest.java
@@ -2,6 +2,7 @@ package nextstep.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.ladder.model.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
