+++ src/test/java/nextstep/ladder/PlayerTest.java
@@ -3,6 +3,7 @@ package nextstep.ladder;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import nextstep.ladder.model.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
