+++ src/test/java/nextstep/ladder/PlayersTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import java.util.List;
+import nextstep.ladder.model.Player;
+import nextstep.ladder.model.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
